# tests/backend/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Algo Atlas",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/taxonomy": {
      "get": {
        "summary": "Get Taxonomy",
        "operationId": "get_taxonomy_api_taxonomy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Taxonomy Api Taxonomy Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Custom Taxonomy",
        "operationId": "create_custom_taxonomy_api_taxonomy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomTaxonomyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Custom Taxonomy Api Taxonomy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/taxonomy/{node_id}": {
      "patch": {
        "summary": "Update Taxonomy Aliases",
        "operationId": "update_taxonomy_aliases_api_taxonomy__node_id__patch",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxonomyAliasUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Taxonomy Aliases Api Taxonomy  Node Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/problems": {
      "get": {
        "summary": "List Problems",
        "operationId": "list_problems_api_problems_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Difficulty"
            }
          },
          {
            "name": "main_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Main Id"
            }
          },
          {
            "name": "subtag_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Subtag Id"
            }
          },
          {
            "name": "taxonomy_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Taxonomy Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Problems Api Problems Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Problem",
        "operationId": "create_problem_api_problems_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Problem Api Problems Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/problems/{problem_id}": {
      "get": {
        "summary": "Get Problem",
        "operationId": "get_problem_api_problems__problem_id__get",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Problem Api Problems  Problem Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Problem",
        "operationId": "update_problem_api_problems__problem_id__patch",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Problem Api Problems  Problem Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Problem",
        "operationId": "delete_problem_api_problems__problem_id__delete",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Problem Api Problems  Problem Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/problems/{problem_id}/mistakes": {
      "post": {
        "summary": "Add Mistake",
        "operationId": "add_mistake_api_problems__problem_id__mistakes_post",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MistakeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Mistake Api Problems  Problem Id  Mistakes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/restore": {
      "post": {
        "summary": "Restore",
        "operationId": "restore_api_export_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Restore Api Export Restore Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/git": {
      "get": {
        "summary": "Get Git State",
        "operationId": "get_git_state_api_git_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Git State Api Git Get"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Git Settings",
        "operationId": "update_git_settings_api_git_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Update Git Settings Api Git Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CustomTaxonomyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Aliases"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CustomTaxonomyCreate"
      },
      "GitSettingsUpdate": {
        "properties": {
          "remote_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Url"
          },
          "branch": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._/-]+$",
            "title": "Branch",
            "default": "main"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          }
        },
        "type": "object",
        "title": "GitSettingsUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MistakeCreate": {
        "properties": {
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "reason_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Reason Ids"
          },
          "observation": {
            "type": "string",
            "maxLength": 2000,
            "title": "Observation",
            "default": ""
          }
        },
        "type": "object",
        "title": "MistakeCreate"
      },
      "ProblemCreate": {
        "properties": {
          "record_initial_mistake": {
            "type": "boolean",
            "title": "Record Initial Mistake",
            "default": true
          },
          "title": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "source": {
            "type": "string",
            "maxLength": 32,
            "title": "Source",
            "default": "leetcode"
          },
          "source_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ],
            "title": "Difficulty",
            "default": "Medium"
          },
          "status": {
            "type": "string",
            "enum": [
              "Open",
              "Understood",
              "Resolved"
            ],
            "title": "Status",
            "default": "Open"
          },
          "primary_subtag_id": {
            "type": "string",
            "title": "Primary Subtag Id"
          },
          "taxonomy_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Taxonomy Ids"
          },
          "failure_reason_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Failure Reason Ids"
          },
          "python_code": {
            "type": "string",
            "maxLength": 200000,
            "title": "Python Code",
            "default": ""
          },
          "time_complexity": {
            "type": "string",
            "maxLength": 80,
            "title": "Time Complexity",
            "default": ""
          },
          "space_complexity": {
            "type": "string",
            "maxLength": 80,
            "title": "Space Complexity",
            "default": ""
          },
          "notes": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Notes"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "observation": {
            "type": "string",
            "maxLength": 2000,
            "title": "Observation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "primary_subtag_id"
        ],
        "title": "ProblemCreate"
      },
      "ProblemUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Easy",
                  "Medium",
                  "Hard"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Open",
                  "Understood",
                  "Resolved"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "primary_subtag_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Subtag Id"
          },
          "taxonomy_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxonomy Ids"
          },
          "python_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000
              },
              {
                "type": "null"
              }
            ],
            "title": "Python Code"
          },
          "time_complexity": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Complexity"
          },
          "space_complexity": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Complexity"
          },
          "notes": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ProblemUpdate"
      },
      "RestoreRequest": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          },
          "review_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Version"
          },
          "decisions": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "keep_local",
                "use_incoming"
              ]
            },
            "type": "object",
            "maxProperties": 500,
            "title": "Decisions"
          }
        },
        "type": "object",
        "title": "RestoreRequest"
      },
      "TaxonomyAliasUpdate": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Aliases"
          }
        },
        "type": "object",
        "title": "TaxonomyAliasUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}